# userServices/src/main/resources/swagger/UserServices.yml
openapi: 3.0.0
info:
  title: User Service API
  version: 1.0.0
  description: API for user profile management, including user login, registration, and deletion.
servers:
  - url: https://api.example.com/user
    description: User Service
paths:
  /login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: string
  /users/{id}:
    delete:
      summary: Delete a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: string
  /register/accRegister:
    post:
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAccRegisterResponse'
      responses:
        '200':
          description: Registration Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAccRegisterResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterAccRegisterResponse:
      type: object
      properties:
        uid:
          type: string
          format: uuid
        firstname:
          type: string
        lastname:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - uid
        - username
        - email
        - password
security:
  - bearerAuth: []